♻️ Refactor: Community view/viewModel 분리
