+++ EatPic-iOS/Sources/Screens/Community/CommunityMainView.swift
@@ -10,10 +10,12 @@ import SwiftUI
 struct CommunityMainView: View {
     
     @EnvironmentObject private var container: DIContainer
-    @State private var selectedUser: CommunityUser = sampleUsers[0]
-    @Bindable private var toastVM = ToastViewModel()
-    @State private var isShowingReportBottomSheet = false
-    @State private var isShowingCommentBottomSheet: Bool = false
+    @State private var viewModel = CommunityMainViewModel()
+    
+//    @State private var selectedUser: CommunityUser = sampleUsers[0]
+//    @Bindable private var toastVM = ToastViewModel()
+//    @State private var isShowingReportBottomSheet = false
+//    @State private var isShowingCommentBottomSheet: Bool = false
     
     var body: some View {
         ScrollView {
@@ -24,28 +26,31 @@ struct CommunityMainView: View {
             }
         }
         .scrollIndicators(.hidden)
-        .toastView(viewModel: toastVM)
+        .toastView(viewModel: viewModel.toastVM)
         .padding(.horizontal, 16)
-        .sheet(isPresented: $isShowingReportBottomSheet) {
-            reportBottomSheetView()
+        .sheet(isPresented: $viewModel.isShowingReportBottomSheet) {
+            ReportBottomSheetView(
+                isShowing: $viewModel.isShowingReportBottomSheet,
+                onReport: viewModel.handleReport
+            )
                 .presentationDetents([.large, .fraction(0.7)])
                 .presentationDragIndicator(.hidden)
         }
-        .sheet(isPresented: $isShowingCommentBottomSheet) {
-            commentBottomSheetView()
+        .sheet(isPresented: $viewModel.isShowingCommentBottomSheet) {
+            CommentBottomSheetView(isShowing: $viewModel.isShowingCommentBottomSheet)
                 .presentationDetents([.large, .fraction(0.7)])
                 .presentationDragIndicator(.hidden)
         }
     }
     
     // 필터링된 카드 리스트
-    private var filteredCards: [PicCard] {
-        if selectedUser.nickname == "전체" {
-            return sampleCards
-        } else {
-            return sampleCards.filter { $0.user == selectedUser }
-        }
-    }
+//    private var filteredCards: [PicCard] {
+//        if selectedUser.nickname == "전체" {
+//            return sampleCards
+//        } else {
+//            return sampleCards.filter { $0.user == selectedUser }
+//        }
+//    }
     
     private func userListView() -> some View {
         ScrollView(.horizontal) {
@@ -55,7 +60,7 @@ struct CommunityMainView: View {
                         ProfileImageView(
                             image: user.profileImage,
                             size: 64,
-                            borderColor: user == selectedUser ? .pink050 : .gray040,
+                            borderColor: user == viewModel.selectedUser ? .pink050 : .gray040,
                             borderWidth: 3
                         )
                         Text(user.id)
@@ -64,7 +69,7 @@ struct CommunityMainView: View {
                     }
                     .padding(EdgeInsets(top: 12, leading: 4, bottom: 0, trailing: 2))
                     .onTapGesture {
-                        selectedUser = user
+                        viewModel.selectUser(user)
                     }
                 }
             }
@@ -76,7 +81,7 @@ struct CommunityMainView: View {
     // 카드 리스트 뷰
     private func cardListView() -> some View {
         LazyVStack(spacing: 32) {
-            ForEach(filteredCards) { card in
+            ForEach(viewModel.filteredCards) { card in
                 // FIXME: - 각 user 당 카드 1개만 프로필 이동 및 메뉴 선택 되는 이슈 (원주연, 25.07.31)
                 PicCardView(
                     profileImage: card.user.profileImage ?? Image(systemName: "person.fill"),
@@ -84,7 +89,7 @@ struct CommunityMainView: View {
                     time: card.time,
                     menuContent: {
                         Button(role: .destructive, action: {
-                            isShowingReportBottomSheet = true
+                            viewModel.isShowingReportBottomSheet = true
                             print("신고하기")
                         }) {
                             Label("신고하기", systemImage: "exclamationmark.bubble")
@@ -95,17 +100,18 @@ struct CommunityMainView: View {
                     onProfileTap: {
                         container.router.push(.userProfile(user: card.user))
                     },
-                    toastVM: toastVM,
-                    onItemAction: { action in
-                        switch action {
-                        case .bookmark(let isOn):
-                            print("북마크 상태: \(isOn)")
-                        case .comment:
-                            isShowingCommentBottomSheet = true  // 여기서 true로 설정
-                        case .reaction(let selected, let counts):
-                            print("선택된 리액션: \(String(describing: selected)), 리액션 수: \(counts)")
-                        }
-                    }
+                    toastVM: viewModel.toastVM,
+                    onItemAction: viewModel.handleCardAction
+//                        { action in
+//                        switch action {
+//                        case .bookmark(let isOn):
+//                            print("북마크 상태: \(isOn)")
+//                        case .comment:
+//                            isShowingCommentBottomSheet = true  // 여기서 true로 설정
+//                        case .reaction(let selected, let counts):
+//                            print("선택된 리액션: \(String(describing: selected)), 리액션 수: \(counts)")
+//                        }
+//                    }
                 )
             }
         }
@@ -134,147 +140,22 @@ struct CommunityMainView: View {
     }
     
     // 댓글 바텀시트 뷰
-    private func commentBottomSheetView() -> some View {
-        ScrollView {
-            BottomSheetView(
-                title: "댓글",
-                content: {
-                    let sampleComments = [
-                        ("user1", "정말 맛있어 보이네요! 🤤"),
-                        ("user2", "어디서 먹을 수 있나요?"),
-                        ("user3", "레시피 공유해주세요~"),
-                        ("user4", "다음에 저도 가봐야겠어요!"),
-                        ("user5", "바로 저장"),
-                        ("user6", "내일 가봐야지")
-                    ]
-                    LazyVStack(spacing: 0) {
-                        ForEach(Array(sampleComments.enumerated()),
-                                id: \.offset) { index, comment in
-                            commentListView(
-                                userName: comment.0, commentText: comment.1,
-                                isLast: index == sampleComments.count - 1)
-                        }
-                        Spacer()
-                    }
-                }
-            )
-            .padding(.top, 24)
-        }
-        .scrollIndicators(.hidden)
-    }
-    
-    private func commentPostView() -> some View {
-        HStack(alignment: .center, spacing: 16) {
-            // 프로필 이미지
-            Circle()
-                .fill(Color.gray040)
-                .frame(width: 40, height: 40)
-            
-            // 댓글 텍스트 필드
-            TextField("댓글 달기...", text: .constant(""))
-                .font(.system(size: 14, weight: .regular, design: .default))
-        }
-        .background(Color.white)
-        .frame(maxWidth: .infinity)
-        .padding(16)
-    }
-    
-    private func commentListView(userName: String, commentText: String, isLast: Bool) -> some View {
-        HStack {
-            // 프로필 이미지
-            Circle()
-                .fill(Color.gray040)
-                .frame(width: 40, height: 40)
-            
-            VStack(alignment: .leading, spacing: 0) {
-                HStack(spacing: 16) {
-                    Text(userName)
-                        .font(.dsHeadline)
-                        .foregroundStyle(Color.gray080)
-                    
-                    Text("10분 전")
-                        .font(.dsSubhead)
-                        .foregroundStyle(Color.gray060)
-                    
-                    Spacer()
-                }
-                
-                Spacer().frame(height: 2)
-                
-                Text(commentText)
-                    .font(.dsCallout)
-                    .foregroundStyle(Color.gray080)
-                
-                Spacer().frame(height: 4)
-                
-                Text("답글 달기")
-                    .font(.dsFootnote)
-                    .foregroundStyle(Color.gray060)
-            }
-            Spacer()
-        }
-        .padding(.horizontal, 16)
-        .padding(.top, 14)
-        .padding(.bottom, 10)
-    }
-    
-    // 신고 바텀시트 뷰
-    private func reportBottomSheetView() -> some View {
-        ScrollView {
-            BottomSheetView(
-                title: "신고하기",
-                subtitle: {
-                    VStack(spacing: 16) {
-                        Text("해당 Pic 카드를 신고하는 이유")
-                            .font(.dsTitle2)
-                            .foregroundStyle(Color.gray080)
-                        Text("회원님의 신고는 익명으로 처리됩니다")
-                            .font(.dsFootnote)
-                            .foregroundStyle(Color.gray060)
-                    }
-                },
-                content: {
-                    let reportTypes = [
-                        "욕설 또는 비방",
-                        "음란성/선정적 내용",
-                        "도배 또는 광고성 게시물",
-                        "거짓 정보 또는 허위 사실",
-                        "불쾌감을 주는 이미지 또는 언행",
-                        "저작권 침해"
-                    ]
-                    
-                    VStack(spacing: 0) {
-                        ForEach(reportTypes, id: \.self) { reportType in
-                            reportListView(reportType: reportType)
-                        }
-                    }
-                })
-            .padding(.top, 24)
-        }
-        .scrollIndicators(.hidden)
-    }
-    
-    private func reportListView(reportType: String) -> some View {
-        VStack(spacing: 0) {
-            Divider().foregroundStyle(Color.gray030)
-            HStack {
-                Text(reportType)
-                    .font(.dsBody)
-                    .foregroundColor(.black)
-                Spacer()
-                Image(systemName: "chevron.right")
-                    .foregroundColor(.gray050)
-            }
-            .padding(.top, 20)
-            .padding(.bottom, 16)
-            .padding(.leading, 28)
-            .padding(.trailing, 16)
-            .contentShape(Rectangle())
-            .onTapGesture {
-                handleReport(reportType)
-            }
-        }
-    }
+//    
+//    private func commentPostView() -> some View {
+//        HStack(alignment: .center, spacing: 16) {
+//            // 프로필 이미지
+//            Circle()
+//                .fill(Color.gray040)
+//                .frame(width: 40, height: 40)
+//            
+//            // 댓글 텍스트 필드
+//            TextField("댓글 달기...", text: .constant(""))
+//                .font(.system(size: 14, weight: .regular, design: .default))
+//        }
+//        .background(Color.white)
+//        .frame(maxWidth: .infinity)
+//        .padding(16)
+//    }
     
     // 신고 버튼 컴포넌트
     private func reportButton(title: String, action: @escaping () -> Void) -> some View {
@@ -297,12 +178,12 @@ struct CommunityMainView: View {
         }
     }
     
-    // 신고 처리 함수
-    private func handleReport(_ reportType: String) {
-        print("신고 유형: \(reportType)")
-        isShowingReportBottomSheet = false
-        toastVM.showToast(title: "신고되었습니다.")
-    }
+//    // 신고 처리 함수
+//    private func handleReport(_ reportType: String) {
+//        print("신고 유형: \(reportType)")
+//        isShowingReportBottomSheet = false
+//        toastVM.showToast(title: "신고되었습니다.")
+//    }
 }
 
 #Preview {
